feat(Video): replace video logic to the component;
