+++ src/components/pages/course/lesson/Lessons.vue
@@ -6,15 +6,18 @@
           <div
             class="relative block min-h-[540px] bg-white border border-gray-200 rounded-lg shadow tablet:min-h-[200px] dark:bg-gray-800 dark:border-gray-700">
             <div class="p-4">
-              <video
-                id="selected-video"
+              <VideoPlayer
                 ref="selectedVideo"
-                class="w-full max-w-full z-10"
-                :poster="posterPreviewLink"
-                playsinline
-                controls
-                :tabindex="tabbingIndex"
-                @keydown="changeVideoSpeed" />
+                :is-available="isAvailable"
+                :current-lesson="currentLesson"
+                @not-found="setNotFound"
+              >
+                <template #default="{ speedChanged, speed }">
+                  <LockedLayer v-if="!unlocked" />
+                  <NotFoundLayer v-if="notFound" />
+                  <PlaybackSpeedLayer :show="speedChanged" :speed="speed" />
+                </template>
+              </VideoPlayer>
               <div class="mt-2 flex justify-between items-center">
                 <p class="basis-4/6 mt-1 mobile:hidden">
                   To change the speed of a video using the keyboard, you can use
@@ -29,9 +32,6 @@
                 </button>
               </div>
             </div>
-            <LockedLayer v-if="!unlocked" />
-            <NotFoundLayer v-if="notFound" />
-            <PlaybackSpeedLayer :show="speedChanged" :speed="speed" />
           </div>
         </div>
         <ul
@@ -53,21 +53,14 @@ import type { PropType } from 'vue';
 import { defineComponent, ref } from 'vue';
 import type { ICourse } from '@/types/ICourse.types';
 import LessonItem from '@/components/pages/course/lesson/LessonItem.vue';
-import Hls from 'hls.js'; 
 import NotFoundLayer from '@/components/pages/course/lesson/VideoLayers/NotFoundLayer.vue';
 import LockedLayer from '@/components/pages/course/lesson/VideoLayers/LockedLayer.vue';
 import PlaybackSpeedLayer from '@/components/pages/course/lesson/VideoLayers/PlaybackSpeedLayer.vue';
-
-export const DURATION_TIME_KEY = '--duration-time';
-
-export const ArrowUpKey = 'ArrowUp';
-export const ArrowDownKey = 'ArrowDown';
-
-export const MIN_RATE = 0.25;
-export const MAX_RATE = 5;
+import VideoPlayer from '@/components/pages/course/lesson/VideoPlayer.vue';
 
 export default defineComponent({
   components: {
+    VideoPlayer,
     PlaybackSpeedLayer,
     LockedLayer,
     NotFoundLayer,
@@ -87,11 +80,8 @@ export default defineComponent({
     }
   },
   data: () => ({
-    hls: new Hls(),
     lessonNumber: 1,
     notFound: false,
-    speed: 1,
-    speedChanged: false,
   }),
   computed: {
     isAvailable(): boolean {
@@ -104,45 +94,25 @@ export default defineComponent({
     currentLesson(): ICourse.Lesson {
       return this.lessons[Number(this.lessonNumber) - 1];
     },
-    posterPreviewLink(): string {
-      return `${this.currentLesson.previewImageLink}/lesson-${this.currentLesson.order}.webp`;
-    },
     unlocked(): boolean {
       return this.currentLesson.status === 'unlocked';
     },
     courseLessonKey(): string {
       return `${this.course.id}-selected-lesson`;
     },
-    tabbingIndex(): number {
-      return !this.isAvailable ? -1 : 0;
-    },
   },
   watch: {
-    currentLesson() {
-      this.loadHlsPlayer();
-    },
     lessonNumber(newValue: number) {
       this.setSelectedLocal(newValue);
-    },
-    speed(newValue: number) {
-      this.speedChanged = true;
-
-      const video = this.selectedVideo as HTMLMediaElement;
-      video.playbackRate = newValue;
-
-      setTimeout(() => (this.speedChanged = false), 400);
     },
   },
   mounted() {
     this.initSelectedLesson();
-    this.loadHlsPlayer();
-  },
-  beforeUnmount() {
-    if (this.hls) {
-      this.hls.destroy();
-    }
   },
   methods: {
+    setNotFound(newValue: boolean) {
+      this.notFound = newValue
+    },
     togglePictureInPicture() {
       if (document.pictureInPictureElement) {
         document.exitPictureInPicture();
@@ -151,80 +121,13 @@ export default defineComponent({
         video.requestPictureInPicture();
       }
     },
-    changeVideoSpeed(event: KeyboardEvent) {
-      if (this.speed <= MIN_RATE || this.speed >= MAX_RATE) {
-        return;
-      }
-      switch (event.key) {
-        case ArrowUpKey:
-          this.speed += MIN_RATE;
-          break;
-        case ArrowDownKey:
-          this.speed -= MIN_RATE;
-          break;
-        default:
-          break;
-      }
-    },
     initSelectedLesson() {
       this.lessonNumber =
         Number(window.localStorage.getItem(this.courseLessonKey)) || 1;
     },
     setSelectedLocal(order: number) {
       window.localStorage.setItem(this.courseLessonKey, order.toString());
-    },
-    checkCurrentTime(): number {
-      const lastSecond = localStorage.getItem(
-        this.currentLesson.id + DURATION_TIME_KEY,
-      );
-      if (lastSecond) {
-        return Number(lastSecond);
-      }
-
-      return 0;
-    },
-    loadHlsPlayer() {
-      const video = this.selectedVideo as HTMLMediaElement;
-
-      this.initVideoListener(video);
-
-      if (!this.isAvailable) {
-        this.resetVideoPlayer();
-        return;
-      }
-
-      try {
-        if (Hls.isSupported()) {
-          this.setHlsSourceAndListener(video);
-          return;
-        }
-      } catch (err) {
-        this.notFound = true;
-      }
-    },
-    initVideoListener(video: HTMLMediaElement) {
-      video.addEventListener('timeupdate', () => {
-        const currentTime = JSON.stringify(video.currentTime);
-        if (!Number(currentTime)) {
-          return;
-        }
-        localStorage.setItem(
-          this.currentLesson.id + DURATION_TIME_KEY,
-          currentTime,
-        );
-      });
-    },
-    resetVideoPlayer() {
-      this.hls.destroy();
-      this.hls = new Hls();
-    },
-    setHlsSourceAndListener(video: HTMLMediaElement) {
-      this.hls.loadSource(this.currentLesson.link);
-      this.hls.attachMedia(video);
-      this.hls.on(Hls.Events.MANIFEST_PARSED, () => {
-        video.currentTime = this.checkCurrentTime();
-      });
-    },
+    }
   },
 });
 </script>
